# frontend/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AffectedStudent" : {
        "properties" : {
          "attendanceId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "studentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "studentName" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "attendanceId", "studentId" ],
        "type" : "object"
      },
      "AgendaView" : {
        "properties" : {
          "events" : {
            "items" : {
              "$ref" : "#/components/schemas/EventView"
            },
            "type" : "array"
          },
          "freeBlocks" : {
            "items" : {
              "$ref" : "#/components/schemas/SlotView"
            },
            "type" : "array"
          }
        },
        "required" : [ "events", "freeBlocks" ],
        "type" : "object"
      },
      "AttendanceView" : {
        "properties" : {
          "canCancel" : {
            "type" : "boolean"
          },
          "canMark" : {
            "type" : "boolean"
          },
          "cancelReason" : {
            "type" : [ "string", "null" ]
          },
          "capacity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "confirmationMethod" : {
            "enum" : [ "QR", "LATER" ],
            "type" : [ "string", "null" ]
          },
          "confirmedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "cycleId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "eventId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "modality" : {
            "enum" : [ "PERSONALIZED", "SEMI_PERSONALIZED" ],
            "type" : "string"
          },
          "occupied" : {
            "format" : "int32",
            "type" : "integer"
          },
          "onlyMarkedByCoach" : {
            "type" : "boolean"
          },
          "override" : {
            "type" : "boolean"
          },
          "overrideReason" : {
            "type" : [ "string", "null" ]
          },
          "rescheduledFrom" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "SCHEDULED", "ATTENDED", "CANCELLED_ON_TIME", "RESCHEDULED", "NO_SHOW", "CANCELLED_BY_COACH" ],
            "type" : "string"
          },
          "studentConfirmed" : {
            "type" : "boolean"
          },
          "studentId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "studentName" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "canCancel", "canMark", "capacity", "cycleId", "endsAt", "eventId", "id", "modality", "occupied", "onlyMarkedByCoach", "override", "startsAt", "status", "studentConfirmed" ],
        "type" : "object"
      },
      "AttendeeView" : {
        "properties" : {
          "attendanceId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "canMark" : {
            "type" : "boolean"
          },
          "confirmationMethod" : {
            "enum" : [ "QR", "LATER" ],
            "type" : [ "string", "null" ]
          },
          "confirmedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "minor" : {
            "type" : "boolean"
          },
          "onlyMarkedByCoach" : {
            "type" : "boolean"
          },
          "override" : {
            "type" : "boolean"
          },
          "overrideReason" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "enum" : [ "SCHEDULED", "ATTENDED", "CANCELLED_ON_TIME", "RESCHEDULED", "NO_SHOW", "CANCELLED_BY_COACH" ],
            "type" : "string"
          },
          "studentConfirmed" : {
            "type" : "boolean"
          },
          "studentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "studentName" : {
            "type" : "string"
          }
        },
        "required" : [ "attendanceId", "canMark", "minor", "onlyMarkedByCoach", "override", "status", "studentConfirmed", "studentId", "studentName" ],
        "type" : "object"
      },
...
      "BlockCreated" : {
        "properties" : {
          "affectedEvents" : {
            "items" : {
              "$ref" : "#/components/schemas/EventView"
            },
            "type" : "array"
          },
          "block" : {
            "$ref" : "#/components/schemas/BlockSummary"
          }
        },
        "required" : [ "affectedEvents", "block" ],
        "type" : "object"
      },
      "BlockInput" : {
        "properties" : {
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reason" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "endsAt", "startsAt" ],
        "type" : "object"
      },
      "BlockSummary" : {
        "properties" : {
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reason" : {
            "type" : [ "string", "null" ]
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "endsAt", "id", "startsAt" ],
        "type" : "object"
      },
...
      "CancelAttendancesCommand" : {
        "properties" : {
          "attendanceIds" : {
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "maxItems" : 50,
            "minItems" : 0,
            "type" : "array"
          },
          "reason" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "attendanceIds", "reason" ],
        "type" : "object"
      },
      "CancelAttendancesResult" : {
        "properties" : {
          "cancelled" : {
            "items" : {
              "$ref" : "#/components/schemas/AttendanceView"
            },
            "type" : "array"
          }
        },
        "required" : [ "cancelled" ],
        "type" : "object"
      },
      "CancelEventCommand" : {
        "properties" : {
          "reason" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "reason" ],
        "type" : "object"
      },
      "CancelResult" : {
        "properties" : {
          "cancelled" : {
            "$ref" : "#/components/schemas/AttendanceView"
          },
          "replacement" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/AttendanceView"
            }, {
              "type" : "null"
            } ]
          }
        },
        "required" : [ "cancelled" ],
        "type" : "object"
      },
      "ChangeCapacityCommand" : {
        "properties" : {
          "capacity" : {
            "format" : "int32",
            "maximum" : 10,
            "minimum" : 2,
            "type" : "integer"
          }
        },
        "required" : [ "capacity" ],
        "type" : "object"
      },
...
      "CoachBookCommand" : {
        "properties" : {
          "override" : {
            "type" : "boolean"
          },
          "overrideReason" : {
            "type" : "string"
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "startsAt" ],
        "type" : "object"
      },
      "CoachCancelCommand" : {
        "properties" : {
          "newStartsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "override" : {
            "type" : "boolean"
          },
          "overrideReason" : {
            "type" : "string"
          },
          "reason" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "reason" ],
        "type" : "object"
      },
...
      "EventCancelResult" : {
        "properties" : {
          "affectedStudents" : {
            "items" : {
              "$ref" : "#/components/schemas/AffectedStudent"
            },
            "type" : "array"
          },
          "event" : {
            "$ref" : "#/components/schemas/EventView"
          }
        },
        "required" : [ "affectedStudents", "event" ],
        "type" : "object"
      },
      "EventView" : {
        "properties" : {
          "attendees" : {
            "items" : {
              "$ref" : "#/components/schemas/AttendeeView"
            },
            "type" : "array"
          },
          "canShowQr" : {
            "type" : "boolean"
          },
          "capacity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "freeSeats" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "modality" : {
            "enum" : [ "PERSONALIZED", "SEMI_PERSONALIZED" ],
            "type" : "string"
          },
          "occupied" : {
            "format" : "int32",
            "type" : "integer"
          },
          "phase" : {
            "enum" : [ "PAST", "NOW", "UPCOMING" ],
            "type" : "string"
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "SCHEDULED", "CANCELLED" ],
            "type" : "string"
          }
        },
        "required" : [ "attendees", "canShowQr", "capacity", "endsAt", "freeSeats", "id", "modality", "occupied", "phase", "startsAt", "status" ],
        "type" : "object"
      },
...
      "MarkCommand" : {
        "properties" : {
          "result" : {
            "enum" : [ "SCHEDULED", "ATTENDED", "CANCELLED_ON_TIME", "RESCHEDULED", "NO_SHOW", "CANCELLED_BY_COACH" ],
            "type" : "string"
          }
        },
        "required" : [ "result" ],
        "type" : "object"
      },
      "MarkEventCommand" : {
        "properties" : {
          "marks" : {
            "items" : {
              "$ref" : "#/components/schemas/MarkItem"
            },
            "minItems" : 1,
            "type" : "array"
          }
        },
        "required" : [ "marks" ],
        "type" : "object"
      },
      "MarkItem" : {
        "properties" : {
          "attendanceId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "SCHEDULED", "ATTENDED", "CANCELLED_ON_TIME", "RESCHEDULED", "NO_SHOW", "CANCELLED_BY_COACH" ],
            "type" : "string"
          }
        },
        "required" : [ "attendanceId", "status" ],
        "type" : "object"
      },
...
      "SlotView" : {
        "properties" : {
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "localDate" : {
            "type" : "string"
          },
          "localTime" : {
            "type" : "string"
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "endsAt", "localDate", "localTime", "startsAt" ],
        "type" : "object"
      },
...
      "WindowInput" : {
        "properties" : {
          "dayOfWeek" : {
            "format" : "int32",
            "maximum" : 7,
            "minimum" : 1,
            "type" : "integer"
          },
          "end" : {
            "minLength" : 1,
            "type" : "string"
          },
          "start" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "dayOfWeek", "end", "start" ],
        "type" : "object"
      },
      "WindowView" : {
        "properties" : {
          "dayOfWeek" : {
            "format" : "int32",
            "type" : "integer"
          },
          "end" : {
            "type" : "string"
          },
          "start" : {
            "type" : "string"
          }
        },
        "required" : [ "dayOfWeek", "end", "start" ],
        "type" : "object"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/coach/agenda" : {
      "get" : {
        "operationId" : "agenda",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AgendaView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
...
    "/api/coach/attendances/{id}/cancel" : {
      "post" : {
        "operationId" : "cancelAttendance",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CoachCancelCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CancelResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
    "/api/coach/attendances/{id}/mark" : {
      "post" : {
        "operationId" : "mark",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MarkCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttendanceView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
    "/api/coach/availability" : {
      "get" : {
        "operationId" : "weekly",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/WindowView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      },
      "put" : {
        "operationId" : "replaceWeekly",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/WindowInput"
                },
                "type" : "array"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/WindowView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
    "/api/coach/availability/blocks" : {
      "get" : {
        "operationId" : "blocks",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/BlockSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      },
      "post" : {
        "operationId" : "createBlock",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BlockInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BlockCreated"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
    "/api/coach/availability/blocks/{id}" : {
      "delete" : {
        "operationId" : "deleteBlock",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
...
    "/api/coach/events/{id}/cancel" : {
      "post" : {
        "operationId" : "cancelEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CancelEventCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventCancelResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
    "/api/coach/events/{id}/capacity" : {
      "put" : {
        "operationId" : "changeCapacity",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangeCapacityCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
    "/api/coach/events/{id}/mark" : {
      "post" : {
        "operationId" : "markEvent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MarkEventCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AttendanceView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
...
    "/api/coach/students/{studentId}/attendances/cancel" : {
      "post" : {
        "operationId" : "cancelAttendances",
        "parameters" : [ {
          "in" : "path",
          "name" : "studentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CancelAttendancesCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CancelAttendancesResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
...
    "/api/coach/students/{studentId}/sessions" : {
      "get" : {
        "operationId" : "attendancesOfStudent",
        "parameters" : [ {
          "in" : "path",
          "name" : "studentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AttendanceView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "scheduling-controller" ]
      },
      "post" : {
        "operationId" : "book_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "studentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CoachBookCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AttendanceView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "scheduling-controller" ]
      }
    },
...
  }
}